+++ tokio-postgres/src/connect_tls.rs
@@ -1,4 +1,4 @@
-use crate::config::SslMode;
+use crate::config::{SslMode, SslNegotiation};
 use crate::maybe_tls_stream::MaybeTlsStream;
 use crate::tls::private::ForcePrivateApi;
 use crate::tls::TlsConnect;
@@ -10,6 +10,7 @@ use tokio::io::{AsyncRead, AsyncReadExt, AsyncWrite, AsyncWriteExt};
 pub async fn connect_tls<S, T>(
     mut stream: S,
     mode: SslMode,
+    negotiation: SslNegotiation,
     tls: T,
     has_hostname: bool,
 ) -> Result<MaybeTlsStream<S, T::Stream>, Error>
@@ -22,22 +23,27 @@ where
         SslMode::Prefer if !tls.can_connect(ForcePrivateApi) => {
             return Ok(MaybeTlsStream::Raw(stream))
         }
+        SslMode::Prefer if negotiation == SslNegotiation::Direct => {
+            return Err(Error::tls("weak sslmode \"prefer\" may not be used with sslnegotiation=direct (use \"require\", \"verify-ca\", or \"verify-full\")".into()))
+        }
         SslMode::Prefer | SslMode::Require | SslMode::VerifyCa | SslMode::VerifyFull => {}
     }
 
-    let mut buf = BytesMut::new();
-    frontend::ssl_request(&mut buf);
-    stream.write_all(&buf).await.map_err(Error::io)?;
+    if negotiation == SslNegotiation::Postgres {
+        let mut buf = BytesMut::new();
+        frontend::ssl_request(&mut buf);
+        stream.write_all(&buf).await.map_err(Error::io)?;
 
-    let mut buf = [0];
-    stream.read_exact(&mut buf).await.map_err(Error::io)?;
+        let mut buf = [0];
+        stream.read_exact(&mut buf).await.map_err(Error::io)?;
 
-    if buf[0] != b'S' {
-        match mode {
-            SslMode::Require | SslMode::VerifyCa | SslMode::VerifyFull => {
-                return Err(Error::tls("server does not support TLS".into()))
+        if buf[0] != b'S' {
+            match mode {
+                SslMode::Require | SslMode::VerifyCa | SslMode::VerifyFull => {
+                    return Err(Error::tls("server does not support TLS".into()))
+                }
+                SslMode::Disable | SslMode::Prefer => return Ok(MaybeTlsStream::Raw(stream)),
             }
-            SslMode::Disable | SslMode::Prefer => return Ok(MaybeTlsStream::Raw(stream)),
         }
     }