+++ postgres-types/CHANGELOG.md
@@ -2,6 +2,22 @@
 
 ## Unreleased
 
+## v0.2.9 - 2025-02-02
+
+### Added
+
+* Added support for `cidr` 0.3 via the `with-cidr-0_3` feature.
+
+### Fixed
+
+* Fixed deserialization of out of bounds inputs to `time` 0.3 types to return an error rather than panic.
+
+## v0.2.8 - 2024-09-15
+
+### Added
+
+* Added support for `jiff` 0.1 via the `with-jiff-01` feature.
+
 ## v0.2.7 - 2024-07-21
 
 ### Added